src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
package ru.yandex.practicum.filmorate.repository;

import ru.yandex.practicum.filmorate.exception.UserNotFoundException;
import ru.yandex.practicum.filmorate.model.User;

import java.util.*;

public class UserRepository {

    private final Map<Long, User> users;

    private long id;

    public UserRepository() {
        users = new HashMap<>();
        id = 0;
    }

    private long generateId() {
        return ++id;
    }

    public List<User> getUsers() {
        return new ArrayList<>(users.values());
    }

    public User addUser(User user) {
        if (user.getName() == null || user.getName().isBlank()) {
            user.setName(user.getLogin());
        }
        user.setId(generateId());
        users.put(user.getId(), user);
        return users.get(user.getId());
    }

    public User updateUser(User user) {
        if (!users.containsKey(user.getId())) {
            throw new UserNotFoundException("Пользователя с id = " + user.getId() + " не существует.");
        }
        if (user.getName() == null || user.getName().isBlank()) {
            user.setName(user.getLogin());
        }
        users.put(user.getId(), user);
        return user;
    }
}
